[chore] move postTweet to WalletComponent
