+++ src/pages/ReceiveTokenPage/components/ReceiveComplete.js
@@ -8,7 +8,6 @@ import Lottie from "react-lottie-player";
 import { useRecoilValue } from "recoil";
 import styled from "styled-components";
 import { getTrxsLinkInfo } from "utils/api/trxs";
-import { postTweet } from "utils/api/twitter";
 import { receiveTrxHashState } from "utils/atoms/trxs";
 import { COLORS as palette } from "utils/style/Color/colors";
 import Typography from "utils/style/Typography/index";
@@ -121,73 +120,6 @@ const ReceiveComplete = ({ walletList, select }) => {
   const [receiveInfo, setReceiveInfo] = useState(null);
   const receiveTrxHash = useRecoilValue(receiveTrxHashState);
 
-  const convertDateFormat = (date) => {
-    const dateArr = date.substring(4, 33).split(" ");
-    let monthNamesEn = {
-      Jan: "January",
-      Feb: "February",
-      Mar: "March",
-      Apr: "April",
-      May: "May",
-      Jun: "June",
-      Jul: "July",
-      Aug: "August",
-      Sep: "September",
-      Oct: "October",
-      Nov: "November",
-      Dec: "December",
-    };
-
-    let monthNamesKo = {
-      Jan: "1",
-      Feb: "2",
-      Mar: "3",
-      Apr: "4",
-      May: "5",
-      Jun: "6",
-      Jul: "7",
-      Aug: "8",
-      Sep: "9",
-      Oct: "10",
-      Nov: "11",
-      Dec: "12",
-    };
-
-    if (localStorage.getItem("language") === "en") {
-      return (
-        dateArr[3].substring(0, 5) +
-        " on " +
-        monthNamesEn[date[0]] +
-        " " +
-        dateArr[1] +
-        ", " +
-        dateArr[2] +
-        " (" +
-        dateArr[4].substring(0, 6) +
-        ":" +
-        dateArr[4].substring(6, 8) +
-        ")" +
-        "\n"
-      );
-    } else {
-      return (
-        dateArr[2] +
-        "년 " +
-        monthNamesKo[dateArr[0]] +
-        "월 " +
-        dateArr[1] +
-        "일 " +
-        dateArr[3].substring(0, 5) +
-        " (" +
-        dateArr[4].substring(0, 6) +
-        ":" +
-        dateArr[4].substring(6, 8) +
-        ")" +
-        "\n"
-      );
-    }
-  };
-
   const txHashExplorerOnClick = () => {
     if (Number(receiveInfo.networkId) == 5) {
       window.open(`https://goerli.etherscan.io/tx/${receiveTrxHash}`);
@@ -195,21 +127,6 @@ const ReceiveComplete = ({ walletList, select }) => {
       window.open(`https://polygonscan.com/tx/${receiveTrxHash}`);
     }
   };
-
-  const pathname = window.location.pathname.split("/");
-  const linkKey = pathname[pathname.length - 1];
-  const requestPostTweet = async () => {
-    const date = new Date().toString();
-    const dateInFormat = convertDateFormat(date);
-
-    await postTweet(linkKey, dateInFormat).then((data) => {
-      console.log(data);
-    });
-  };
-
-  useEffect(() => {
-    requestPostTweet();
-  }, []);
 
   useEffect(() => {
     const pathname = window.location.pathname.split("/");
